# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leakr API - connecting journalists with anonymous sources
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type LeakSensitivity =
  (typeof LeakSensitivity)[keyof typeof LeakSensitivity];

export const LeakSensitivity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type LeakStatus = (typeof LeakStatus)[keyof typeof LeakStatus];

export const LeakStatus = {
  pending: "pending",
  verified: "verified",
  claimed: "claimed",
  published: "published",
} as const;

export interface Leak {
  id: number;
  title: string;
  teaser: string;
  category: string;
  sensitivity: LeakSensitivity;
  status: LeakStatus;
  anonymousHandle: string;
  documentCount: number;
  /** @nullable */
  claimedByJournalistId: number | null;
  /** @nullable */
  claimedByJournalistName: string | null;
  viewCount: number;
  /** @nullable */
  verifiedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateLeakBodySensitivity =
  (typeof CreateLeakBodySensitivity)[keyof typeof CreateLeakBodySensitivity];

export const CreateLeakBodySensitivity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateLeakBody {
  title: string;
  teaser: string;
  category: string;
  sensitivity: CreateLeakBodySensitivity;
  documentCount: number;
}

export type UpdateLeakBodyStatus =
  (typeof UpdateLeakBodyStatus)[keyof typeof UpdateLeakBodyStatus];

export const UpdateLeakBodyStatus = {
  pending: "pending",
  verified: "verified",
  claimed: "claimed",
  published: "published",
} as const;

export interface UpdateLeakBody {
  status?: UpdateLeakBodyStatus;
  /** @nullable */
  verifiedAt?: string | null;
}

export interface ClaimLeakBody {
  journalistId: number;
}

export interface Journalist {
  id: number;
  displayName: string;
  outlet: string;
  bio: string;
  specializations: string[];
  verificationBadge: boolean;
  leaksClaimed: number;
  leaksPublished: number;
  avatarInitials: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateJournalistBody {
  displayName: string;
  outlet: string;
  bio: string;
  specializations: string[];
}

export interface UpdateJournalistBody {
  displayName?: string;
  outlet?: string;
  bio?: string;
  specializations?: string[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  leakCount: number;
  iconName: string;
}

export interface PlatformStats {
  totalLeaks: number;
  totalJournalists: number;
  totalCategories: number;
  verifiedLeaks: number;
  claimedLeaks: number;
  publishedLeaks: number;
  pendingLeaks: number;
  criticalLeaks: number;
}

export type ActivityEventEventType =
  (typeof ActivityEventEventType)[keyof typeof ActivityEventEventType];

export const ActivityEventEventType = {
  leak_submitted: "leak_submitted",
  leak_verified: "leak_verified",
  leak_claimed: "leak_claimed",
  leak_published: "leak_published",
} as const;

export type ActivityEventSensitivity =
  (typeof ActivityEventSensitivity)[keyof typeof ActivityEventSensitivity];

export const ActivityEventSensitivity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface ActivityEvent {
  id: number;
  eventType: ActivityEventEventType;
  description: string;
  timestamp: string;
  category: string;
  sensitivity: ActivityEventSensitivity;
}

export type SensitivityCountSensitivity =
  (typeof SensitivityCountSensitivity)[keyof typeof SensitivityCountSensitivity];

export const SensitivityCountSensitivity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface SensitivityCount {
  sensitivity: SensitivityCountSensitivity;
  count: number;
}

export type ListLeaksParams = {
  category?: string;
  status?: ListLeaksStatus;
  sensitivity?: ListLeaksSensitivity;
  search?: string;
};

export type ListLeaksStatus =
  (typeof ListLeaksStatus)[keyof typeof ListLeaksStatus];

export const ListLeaksStatus = {
  pending: "pending",
  verified: "verified",
  claimed: "claimed",
  published: "published",
} as const;

export type ListLeaksSensitivity =
  (typeof ListLeaksSensitivity)[keyof typeof ListLeaksSensitivity];

export const ListLeaksSensitivity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;