# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leakr API - connecting journalists with anonymous sources
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventEventType } from "./activityEventEventType";
import type { ActivityEventSensitivity } from "./activityEventSensitivity";

export interface ActivityEvent {
  id: number;
  eventType: ActivityEventEventType;
  description: string;
  timestamp: string;
  category: string;
  sensitivity: ActivityEventSensitivity;
}