# lib/api-zod/src/generated/types/updateLeakBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leakr API - connecting journalists with anonymous sources
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLeakBodyStatus } from "./updateLeakBodyStatus";

export interface UpdateLeakBody {
  status?: UpdateLeakBodyStatus;
  /** @nullable */
  verifiedAt?: string | null;
}